feat(stats-menu, profile-coins, shop-menu): Made stats menu configurable through config. Properly load and save the coins of a player profile, default amount of coins for a new profile is set to 100. Implemented a basic shop menu -> properly subtract the coins by the price of each purchased product, added a list of owned-products into the Profile class.
TODO: Think of a way to grant coins, maybe each kill or at a certain killstreak value, maybe in future event prizes, because why not. In addition to that, I will need to handle perk selection, so that a player has one selected perk, and not all of them combined.